fix UB, big state check, macros update

+++ src/mem/s_slice.rs
@@ -3,9 +3,8 @@ use crate::mem::allocator::EMPTY_PTR;
 use crate::mem::free_block::FreeBlock;
 use crate::mem::{StablePtr, StablePtrBuf};
 use crate::utils::mem_context::stable;
-use std::usize;
 
-pub(crate) const ALLOCATED: u64 = 2usize.pow(u64::BITS - 1) as u64; // first biggest bit set to 1, other set to 0
+pub(crate) const ALLOCATED: u64 = 2u64.pow(u64::BITS - 1); // first biggest bit set to 1, other set to 0
 pub(crate) const FREE: u64 = ALLOCATED - 1; // first biggest bit set to 0, other set to 1
 
 /// A smart-pointer for stable memory.
